requirements.txt
certifi==2019.9.11
chardet==3.0.4
clearbit==0.1.7
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
django-cors-headers==3.1.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
factory-boy==2.12.0
Faker==2.0.2
idna==2.8
oauthlib==3.1.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.5
